refactor audio go
