+++ internal/audio/audiofile/audiofile.go
@@ -17,7 +17,6 @@ import (
 	"strings"
 	"talkliketv.click/tltv/internal/models"
 	"talkliketv.click/tltv/internal/util"
-	"unicode"
 
 	"github.com/labstack/echo/v4"
 	audio "talkliketv.click/tltv/internal/audio/pattern"
@@ -42,7 +41,7 @@ var AudioPauseFilePath = map[int]string{
 var (
 	ErrOneFile           = errors.New("no need to zip one file")
 	ErrUnableToParseFile = func(err error) error {
-		return fmt.Errorf("unable to parse file: %s", err)
+		return fmt.Errorf("unable to parsefile file: %s", err)
 	}
 	endSentenceMap = map[rune]bool{
 		'!': true,
@@ -51,11 +50,6 @@ var (
 	}
 	// Use a regular expression to match punctuation characters
 	reAlpha = regexp.MustCompile(`[a-zA-Z]`)
-)
-
-const (
-	minimumPhraseLength = 4
-	maximumPhraseLength = 10
 )
 
 type AudioFileX interface {
@@ -90,7 +84,7 @@ func (r *RealCmdRunner) CombinedOutput(cmd *exec.Cmd) ([]byte, error) {
 // GetLines determines if the uploaded file is an srt, in paragraph form, or one phrase per
 // line and then parses the file accordingly, returning a string slice containing the
 // phrases to be translated
-func (a *AudioFile) GetLines(e echo.Context, f multipart.File) ([]string, error) {
+func (af *AudioFile) GetLines(e echo.Context, f multipart.File) ([]string, error) {
 	// get file type, options are srt, single line text or paragraph
 	fileType := ""
 	scanner := bufio.NewScanner(f)
@@ -160,187 +154,17 @@ func (a *AudioFile) GetLines(e echo.Context, f multipart.File) ([]string, error)
 		stringsSlice = parseSingle(f)
 	}
 	if len(stringsSlice) == 0 {
-		return nil, errors.New("unable to parse file")
+		return nil, errors.New("unable to parsefile file")
 	}
 
 	// remove strings longer than 150 characters and duplicates from stringsSlice
 	return util.RemoveLongStr(util.RemoveDuplicateStr(stringsSlice)), nil
 }
 
-// parseSrt takes a srt multipart file and parses it into a slice of strings
-func parseSrt(f multipart.File) []string {
-	var stringsSlice []string
-	scanner := bufio.NewScanner(f)
-	scanner.Scan()
-	var line string
-	for scanner.Scan() {
-		line = strings.TrimSpace(scanner.Text())
-		if line == "" {
-			continue
-		} else if line[0] >= '0' && line[0] <= '9' {
-			continue
-		} else if line[0] == '[' && line[len(line)-1] == ']' {
-			continue
-		} else if strings.Contains(line, "<font") || strings.Contains(line, "font>") {
-			continue
-		} else {
-			// if the next line following subtitle is not new line it is more dialogue so combine it
-			scanner.Scan()
-			nextLine := scanner.Text()
-			if nextLine != "" {
-				line = strings.ReplaceAll(line, "\n", "")
-				line = line + " " + nextLine
-				line = replaceFmt(line)
-			} else {
-				line = replaceFmt(line)
-			}
-		}
-
-		phrases := splitLongPhrases(line)
-		stringsSlice = append(stringsSlice, phrases...)
-	}
-
-	return stringsSlice
-}
-
-func splitLongPhrases(line string) []string {
-	var splitString []string
-
-	words := strings.Fields(line)
-	// if phrase is too short don't keep it
-	if len(words) <= minimumPhraseLength {
-		return []string{}
-	} else if len(words) < maximumPhraseLength {
-		// if phrase isn't too long don't split it
-		return []string{line}
-	} else {
-		// split into an array of strings along punctuation
-		last := 0
-		for i, word := range words {
-			if unicode.IsPunct(rune(word[len(word)-1])) {
-				nextString := ""
-				for j := last; j <= i; j++ {
-					nextString = nextString + words[j] + " "
-				}
-				splitString = append(splitString, nextString)
-				last = i + 1
-			}
-		}
-		// if last word does not end in punctuation add that string
-		if last < len(words) {
-			splitString = append(splitString, strings.Join(words[last:], " "))
-		}
-		// if long phrase has punctuation split on punctuation
-		if len(splitString) > 1 {
-			// combine any strings that are less than the minimumPhraseLength with the string after it
-			i := 0
-			for i < len(splitString)-1 {
-				// if phrase is small combine it with the next one
-				wordsInString := strings.Fields(splitString[i])
-				if len(wordsInString) < minimumPhraseLength {
-					splitString[i] = splitString[i] + " " + splitString[i+1]
-					// remove the next index of split string
-					splitString = append(splitString[:i+1], splitString[i+2:]...)
-				} else {
-					// if both combined are less than maximum than concat
-					next := splitString[i] + " " + splitString[i+1]
-					nextWordCount := strings.Fields(next)
-					if len(nextWordCount) <= maximumPhraseLength {
-						splitString[i] = next
-						splitString = append(splitString[:i+1], splitString[i+2:]...)
-					}
-				}
-				// else continue
-				i++
-			}
-
-			// now check the last index and pen ultimate of the split string and combine if shorter than minimumPhraseLength
-			if len(splitString) > 1 {
-				lastElem := len(splitString) - 1
-				lastElemCount := len(strings.Fields(splitString[lastElem]))
-				penUltimateElemCount := len(strings.Fields(splitString[lastElem-1]))
-				if lastElemCount < minimumPhraseLength || penUltimateElemCount < minimumPhraseLength {
-					lastString := splitString[lastElem-1] + " " + splitString[lastElem]
-					splitString[lastElem] = lastString
-					splitString = splitString[:lastElem-1]
-				}
-			}
-		} else {
-			return []string{line}
-		}
-	}
-
-	for i := range splitString {
-		splitString[i] = strings.ReplaceAll(splitString[i], "  ", " ")
-		splitString[i] = strings.TrimSpace(splitString[i])
-	}
-	return splitString
-}
-
-// parseParagraph takes a txt multipart file in paragraph form and returns a slice of strings
-func parseParagraph(f multipart.File) []string {
-	var stringsSlice []string
-	scanner := bufio.NewScanner(f)
-	for scanner.Scan() {
-		line := scanner.Text()
-		// Split on punctuation characters
-		last := 0
-		for i, c := range line {
-			if i == len(line)-1 {
-				sentence := strings.TrimSpace(line[last : i+1])
-				last = i + 1
-				words := strings.Fields(sentence)
-				if len(words) > 3 {
-					stringsSlice = append(stringsSlice, line)
-				}
-			} else if endSentenceMap[c] {
-				sentence := strings.TrimSpace(line[last : i+1])
-				last = i + 1
-				phrases := splitLongPhrases(sentence)
-				stringsSlice = append(stringsSlice, phrases...)
-			}
-		}
-	}
-
-	return stringsSlice
-}
-
-// parseSingle takes a txt multipart file with one phrase per line and parses it
-// into a slice of strings
-func parseSingle(f multipart.File) []string {
-	var stringsSlice []string
-	scanner := bufio.NewScanner(f)
-	for scanner.Scan() {
-		line := scanner.Text()
-		phrases := splitLongPhrases(line)
-		stringsSlice = append(stringsSlice, phrases...)
-	}
-
-	return stringsSlice
-}
-
-// replaceFmt is a helper function for parseSrt that replaces characters that are not part
-// of the phrase like descriptions or tags
-func replaceFmt(line string) string {
-	// remove any characters between brackets and brackets [...] or {...} or <...>
-	re := regexp.MustCompile("\\[.*?]") //nolint:gosimple
-	line = re.ReplaceAllString(line, "")
-	re = regexp.MustCompile("\\{.*?}") //nolint:gosimple
-	line = re.ReplaceAllString(line, "")
-	re = regexp.MustCompile("<.*?>")
-	line = re.ReplaceAllString(line, "")
-	line = strings.ReplaceAll(line, "-", "")
-	line = strings.ReplaceAll(line, "♪", "")
-	line = strings.ReplaceAll(line, "\"", "")
-	line = strings.TrimSpace(line)
-
-	return line
-}
-
 // CreateMp3Zip takes the input txt files created with BuildAudioInputFiles and uses ffmpeg
 // to build an output mp3's file and the zips them into a single file to be returned to the
 // requester
-func (a *AudioFile) CreateMp3Zip(e echo.Context, t models.Title, tmpDir string) (*os.File, error) {
+func (af *AudioFile) CreateMp3Zip(e echo.Context, t models.Title, tmpDir string) (*os.File, error) {
 	// get a list of files from the temp directory
 	files, err := os.ReadDir(tmpDir)
 	if err != nil {
@@ -363,7 +187,7 @@ func (a *AudioFile) CreateMp3Zip(e echo.Context, t models.Title, tmpDir string) 
 		cmd := exec.Command("ffmpeg", "-f", "concat", "-safe", "0", "-i", tmpDir+f.Name(), "-c", "copy", outputString) //nolint:gosec
 
 		//Execute the command and get the output
-		output, err := a.cmdX.CombinedOutput(cmd)
+		output, err := af.cmdX.CombinedOutput(cmd)
 		if err != nil {
 			e.Logger().Error(err)
 			e.Logger().Error("combined output: " + string(output))
@@ -430,7 +254,7 @@ func addFileToZip(e echo.Context, zipWriter *zip.Writer, filename string) error 
 
 // BuildAudioInputFiles creates a file with the filepaths of the mp3's used to construct
 // the output files with ffmpeg in CreateMp3Zip
-func (a *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause, from, to, tmpDir string) error {
+func (af *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause, from, to, tmpDir string) error {
 	maxP := len(t.TitlePhrases) - 1
 
 	pattern := audio.GetPattern(t.Pattern)
@@ -525,7 +349,7 @@ func (a *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause, 
 // more phrases than the limit of config.MaxNumPhrases. It takes a iter.Seq of strings and outputs them
 // to files, each chunk containing config.MaxNumPhrases and than zips them up. Sending them back to the
 // user
-func (a *AudioFile) CreatePhrasesZip(e echo.Context, chunkedPhrases iter.Seq[[]string], tmpPath, filename string) (*os.File, error) {
+func (af *AudioFile) CreatePhrasesZip(e echo.Context, chunkedPhrases iter.Seq[[]string], tmpPath, filename string) (*os.File, error) {
 	// create outputs folder to hold all the txt files to zip
 	err := os.MkdirAll(tmpPath, 0777)
 	if err != nil {
